4-4,4-5 [refactor] 두 가지 기능 User쪽에서 Admin으로 이동
