+++ src/main/java/com/example/applefarm_/admin/service/AdminService.java
@@ -6,10 +6,14 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 public interface AdminService {
-    @Transactional
+
     List<UserResponseDto> findCustomerList();
 
-    @Transactional
+
     List<UserResponseDto> findSellerList();
+
+    void modifideroleCustomer(Long id);
+
+    void modifideroleSeller(Long id);
 
 }
