+++ src/main/java/com/example/applefarm_/user/controller/UserController.java
@@ -61,13 +61,4 @@ public class UserController {
         return userService.editUserProfile(userProfileRequestDto, userDetails.getUser().getId());
     }
 
-    @PutMapping("/modifideroleCustomer/{id}")
-    public void modifideroleCustomer(@PathVariable Long id){
-        userService.modifideroleCustomer(id);
-    }
-
-    @PutMapping("/modifideroleSeller/{id}")
-    public void modifideroleSeller(@PathVariable Long id){
-        userService.modifideroleSeller(id);
-    }
 }
