+++ src/main/java/com/example/applefarm_/user/service/UserService.java
@@ -26,13 +26,10 @@ public interface UserService {
 
     ResponseEntity getProductList(int page, int size);
 
-    @Transactional
+
     ResponseEntity getSellerList(int page, int size);
 
 
     UserProfileResponseDto editUserProfile(UserProfileRequestDto userProfileRequestDto, Long id);
 
-
-    void modifideroleCustomer(Long id);
-    void modifideroleSeller(Long id);
 }
